fix provider errors and start localizing
